Move Scan printing tests into tests.scan.test_printing
